+++ crates/stackable-operator/src/builder/pod/container.rs
@@ -1,14 +1,26 @@
+use std::fmt;
+
 use k8s_openapi::api::core::v1::{
     ConfigMapKeySelector, Container, ContainerPort, EnvVar, EnvVarSource, Lifecycle,
     LifecycleHandler, ObjectFieldSelector, Probe, ResourceRequirements, SecretKeySelector,
     SecurityContext, VolumeMount,
 };
-use std::fmt;
+use snafu::Snafu;
 
 use crate::{
-    commons::product_image_selection::ResolvedProductImage, error::Error,
-    validation::is_rfc_1123_label,
+    commons::product_image_selection::ResolvedProductImage, validation::is_rfc_1123_label,
 };
+
+type Result<T, E = Error> = std::result::Result<T, E>;
+
+#[derive(Debug, PartialEq, Snafu)]
+pub enum Error {
+    #[snafu(display("container name {container_name:?} is invalid: {violation:?}"))]
+    InvalidContainerName {
+        container_name: String,
+        violation: String,
+    },
+}
 
 /// A builder to build [`Container`] objects.
 ///
@@ -32,7 +44,7 @@ pub struct ContainerBuilder {
 }
 
 impl ContainerBuilder {
-    pub fn new(name: &str) -> Result<Self, Error> {
+    pub fn new(name: &str) -> Result<Self> {
         Self::validate_container_name(name)?;
         Ok(ContainerBuilder {
             name: name.to_string(),
@@ -264,7 +276,7 @@ impl ContainerBuilder {
 
     /// Validates a container name is according to the [RFC 1123](https://www.ietf.org/rfc/rfc1123.txt) standard.
     /// Returns [Ok] if the name is according to the standard, and [Err] if not.
-    fn validate_container_name(name: &str) -> Result<(), Error> {
+    fn validate_container_name(name: &str) -> Result<()> {
         let validation_result = is_rfc_1123_label(name);
 
         match validation_result {
@@ -355,8 +367,8 @@ mod tests {
 
     use super::*;
     use crate::{
-        builder::{
-            pod::container::{ContainerBuilder, ContainerPortBuilder, FieldPathEnvVar},
+        builder::pod::{
+            container::{ContainerBuilder, ContainerPortBuilder, FieldPathEnvVar},
             resources::ResourceRequirementsBuilder,
         },
         commons::resources::ResourceRequirementsType,
@@ -490,15 +502,12 @@ mod tests {
                 panic!("Container name exceeding 63 characters should cause an error");
             }
             Err(error) => match error {
-                crate::error::Error::InvalidContainerName {
+                Error::InvalidContainerName {
                     container_name,
                     violation,
                 } => {
                     assert_eq!(container_name.as_str(), long_container_name);
                     assert_eq!(violation.as_str(), "must be no more than 63 characters")
-                }
-                _ => {
-                    panic!("InvalidContainerName error expected")
                 }
             },
         }
@@ -569,15 +578,12 @@ mod tests {
                 panic!("Container name exceeding 63 characters should cause an error");
             }
             Err(error) => match error {
-                crate::error::Error::InvalidContainerName {
+                Error::InvalidContainerName {
                     container_name: _,
                     violation,
                 } => {
                     println!("{violation}");
                     assert!(violation.contains(expected_err_contains));
-                }
-                _ => {
-                    panic!("InvalidContainerName error expected");
                 }
             },
         }